Migrate definitions
